pyproject.toml
[project]
name = "dacon-smartmh-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # 기본 데이터 처리 및 시각화
    "pandas>=2.3.2",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",

    # 기본 머신러닝
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",

    # 고급 Gradient Boosting 모델들
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "catboost>=1.2.0",

    # 하이퍼파라미터 최적화
    "optuna>=3.4.0",
    "hyperopt>=0.2.7",
    "scikit-optimize>=0.9.0",

    # 모델 해석성
    "shap>=0.42.0",
    "eli5>=0.13.0",
    "lime>=0.2.0",

    # 딥러닝 (TabNet)
    "pytorch-tabnet>=4.1.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # 차원 축소 및 고급 특성
    "umap-learn>=0.5.4",

    # 실험 추적
    "mlflow>=2.8.0",
    "wandb>=0.15.0",

    # 시각화 고급 기능
    "yellowbrick>=1.5.0",

    # AutoML
    "autogluon.tabular>=1.0.0",
    "flaml>=2.1.0",

    # 유틸리티
    "tqdm>=4.66.0",
    "ipywidgets>=8.1.0",
]
